modify, pagination, login add

+++ src/main/webapp/customer/delete.jsp
@@ -1,3 +1,4 @@
+<%@page import="vo.Customer"%>
 <%@page import="dao.CustomerDao"%>
 <%@ page contentType="text/html; charset=utf-8" pageEncoding="utf-8" %>
 <%
@@ -6,6 +7,13 @@
 	
 	// 업무로직 수행
 	CustomerDao dao = new CustomerDao();
+	
+	Customer customer = dao.getCustomerById(id);
+	if ("No".equals(customer.getDisabled())) {
+		response.sendRedirect("detail.jsp?id=" + id + "&err=fail");
+		return;
+	}
+	
 	dao.deleteCustomerById(id);
 	
 	// 재요청 URL
